3.1.2 删除：如果有子节点，不允许删除。刪除直接後台刪除，因為有真實ID了。刪除后，其他是不是跟著刪除，包括圖要不要調整？

+++ TrizUI/CodesTool/ListCtrlCodes.cs
@@ -46,9 +46,11 @@ namespace CodesTool
 
         $scope.Delete = function (ID) {
             bootbox.confirm(""要删除当前的记录？"", function (result) {
-                requestService.delete(Sources, ID).then(function (data) {
-                    Get{BusinessObjectInfo.ObjName}s();
-                });
+                if (result) {
+                    requestService.delete(Sources, ID).then(function (data) {
+                        Get{BusinessObjectInfo.ObjName}s();
+                    });
+                }
             });
         }
